Small Refactoring.
