+++ editor/components/map/editor-state/shaders/dark_overlay.vs
